Move chart drill-down pages to root dir
